Shop Funktionalitäten hinzugefügt.

+++ article/src/main/java/de/leuphana/article/configuration/ArticleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.leuphana.article.component.structure.Article;
 
-public interface ArticleRepository extends JpaRepository<Article, Long>{
+public interface ArticleRepository extends JpaRepository<Article, Integer>{
